+++ ddtrace/appsec/_iast/_taint_tracking/__init__.py
@@ -2,6 +2,7 @@ import os
 from typing import Any
 from typing import Tuple
 
+from ddtrace.internal._unpatched import _threading as threading
 from ddtrace.internal.logger import get_logger
 from ddtrace.internal.utils.formats import asbool
 
@@ -107,8 +108,12 @@ __all__ = [
 ]
 
 
+def _is_iast_debug_enabled():
+    return asbool(os.environ.get(IAST.ENV_DEBUG, "false"))
+
+
 def iast_taint_log_error(msg):
-    if asbool(os.environ.get(IAST.ENV_DEBUG, "false")):
+    if _is_iast_debug_enabled():
         import inspect
 
         stack = inspect.stack()
@@ -175,3 +180,54 @@ def get_tainted_ranges(pyobject: Any) -> Tuple:
     except ValueError as e:
         iast_taint_log_error("Get ranges error (pyobject type %s): %s" % (type(pyobject), e))
     return tuple()
+
+
+if _is_iast_debug_enabled():
+    TAINTED_FRAMES = []
+
+    def trace_calls_and_returns(frame, event, arg):
+        try:
+            co = frame.f_code
+            func_name = co.co_name
+            if func_name == "write":
+                # Ignore write() calls from print statements
+                return
+            if func_name in ("is_pyobject_tainted", "__repr__"):
+                return
+            line_no = frame.f_lineno
+            filename = co.co_filename
+            if "ddtrace" in filename:
+                return
+            if event == "call":
+                f_locals = frame.f_locals
+                if any([is_pyobject_tainted(f_locals[arg]) for arg in f_locals]):
+                    TAINTED_FRAMES.append(frame)
+                    log.debug("Call to %s on line %s of %s, args: %s", func_name, line_no, filename, frame.f_locals)
+                    log.debug("Tainted arguments:")
+                    for arg in f_locals:
+                        if is_pyobject_tainted(f_locals[arg]):
+                            log.debug("\t%s: %s", arg, f_locals[arg])
+                    log.debug("-----")
+
+                return trace_calls_and_returns
+            elif event == "return":
+                if frame in TAINTED_FRAMES:
+                    TAINTED_FRAMES.remove(frame)
+                    log.debug("Return from %s on line %d of %s, return value: %s", func_name, line_no, filename, arg)
+                    if isinstance(arg, (str, bytes, bytearray, list, tuple, dict)):
+                        if (
+                            (isinstance(arg, (str, bytes, bytearray)) and is_pyobject_tainted(arg))
+                            or (isinstance(arg, (list, tuple)) and any([is_pyobject_tainted(x) for x in arg]))
+                            or (isinstance(arg, dict) and any([is_pyobject_tainted(x) for x in arg.values()]))
+                        ):
+                            log.debug("Return value is tainted")
+                        else:
+                            log.debug("Return value is NOT tainted")
+                    log.debug("-----")
+            return
+
+        except Exception:
+            # Ensure that we do not break by any means
+            pass
+
+    threading.settrace(trace_calls_and_returns)
